update clinic source

+++ clinic/doctor-add.php
@@ -4,6 +4,94 @@ include('./includes/path.inc.php');
 include('./includes/session.inc.php');
 
 include(SELECT_HELPER);
+include(EMAIL_HELPER);
+
+$errFName = $errLName = $errSpec = $errYears = $errSpoke = $errGender = $errEmail = $errContact =  "";
+$classFName = $classLName = $classSpec = $classYears = $classSpoke = $classGender = $classEmail = $errContact = "";
+
+if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+    $fname       = escape_input($_POST['inputFirstName']);
+    $lname       = escape_input($_POST['inputLastName']);
+    if (isset($_POST['inputSpeciality'])) {
+        $speciality = escape_input($_POST['inputSpeciality']);
+    }
+    $years      = escape_input($_POST['inputYrsExp']);
+    $desc       = escape_input($_POST['inputDesc']);
+    if (isset($_POST['inputLanguages'])) {
+        $lang = $_POST['inputLanguages'];
+        $spoke = implode(",", $lang);
+    }
+    $dob        = escape_input($_POST['inputDOB']);
+    if (isset($_POST['inputGender'])) {
+        $gender     = escape_input($_POST['inputGender']);
+    }
+    $email      = escape_input($_POST['inputEmailAddress']);
+    $contact    = escape_input($_POST['inputContactNumber']);
+
+    if (empty($fname)) {
+        $errFName = $error_html['errFirstName'];
+        $classFName = $error_html['errClass'];
+    } else {
+        if (!preg_match($regrex['text'], $fname)) {
+            $errFName = $error_html['invalidText'];
+            $classFName = $error_html['errClass'];
+        }
+    }
+
+    if (empty($lname)) {
+        $errLName = $error_html['errLastName'];
+        $classLName = $error_html['errClass'];
+    } else {
+        if (!preg_match($regrex['text'], $lname)) {
+            $errFName = $error_html['invalidText'];
+            $classFName = $error_html['errClass'];
+        }
+    }
+
+    if (empty($speciality)) {
+        $errSpec = $error_html['errSpec'];
+        $classSpec = $error_html['errClass'];
+    }
+
+    if (empty($years)) {
+        $errYears = $error_html['errYears'];
+        $classYears = $error_html['errClass'];
+    } else {
+        if (!filter_var($years, FILTER_VALIDATE_INT)) {
+            $errYears = $error_html['invalidInt'];
+            $classYears = $error_html['errClass'];
+        }
+    }
+
+    if (empty($lang)) {
+        $errSpoke = $error_html['errSpoke'];
+        $classSpoke = $error_html['errClass'];
+    }
+    if (empty($gender)) {
+        $errGender = $error_html['errGender'];
+        $classGender = $error_html['errClass'];
+    }
+
+    if (empty($email)) {
+        $errEmail = $error_html['errEmail'];
+        $classEmail = $error_html['errClass'];
+    } else {
+        if (!filter_var($email, FILTER_VALIDATE_EMAIL)) {
+            $errEmail =  $error_html['invalidEmail'];
+            $classEmail = $error_html['errClass'];
+        }
+    }
+
+    if (empty($contact)) {
+        $errContact = $error_html['errContact'];
+        $classContact = $error_html['errClass'];
+    } else {
+        if (!preg_match($regrex['contact'], $contact)) {
+            $errContact = $error_html['invalidInt'];
+            $classContact = $error_html['errClass'];
+        }
+    }
+}
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -54,17 +142,19 @@ include(SELECT_HELPER);
                                 <div class="form-row">
                                     <div class="form-group col-md-6">
                                         <label for="inputFirstName">First Name</label>
-                                        <input type="text" name="inputFirstName" class="form-control" id="inputFirstName" placeholder="Enter First Name">
+                                        <input type="text" name="inputFirstName" class="form-control <?php echo $classFName ?>" id="inputFirstName" placeholder="Enter First Name">
+                                        <?php echo $errFName; ?>
                                     </div>
                                     <div class="form-group col-md-6">
                                         <label for="inputLastName">Last Name/Surname</label>
-                                        <input type="text" name="inputLastName" class="form-control" id="inputLastName" placeholder="Enter First Name">
+                                        <input type="text" name="inputLastName" class="form-control <?php echo $classFName ?>" id="inputLastName" placeholder="Enter First Name">
+                                        <?php echo $errLName; ?>
                                     </div>
                                 </div>
                                 <div class="form-row">
                                     <div class="form-group col-md-6">
                                         <label for="inputSpeciality">Speciality</label>
-                                        <select name="inputSpeciality" id="inputSpeciality" class="form-control selectpicker" data-live-search="true">
+                                        <select name="inputSpeciality" id="inputSpeciality" class="form-control selectpicker <?= $classSpec ?>" data-live-search="true">
                                             <option value="" selected disabled>Choose</option>
                                             <?php
                                             $table_result = mysqli_query($conn, "SELECT * FROM speciality");
@@ -73,10 +163,12 @@ include(SELECT_HELPER);
                                             }
                                             ?>
                                         </select>
+                                        <?= $errSpec ?>
                                     </div>
                                     <div class="form-group col-md-6">
                                         <label for="inputYrsExp">Year Experience</label>
-                                        <input type="number" name="inputYrsExp" class="form-control" id="inputYrsExp" placeholder="Enter Years Experience">
+                                        <input type="text" name="inputYrsExp" class="form-control <?= $classYears ?>" id="inputYrsExp" placeholder="Enter Years Experience">
+                                        <?= $errYears ?>
                                     </div>
                                 </div>
                                 <!-- End Add Doctor -->
@@ -116,14 +208,16 @@ include(SELECT_HELPER);
                     <div class="card">
                         <div class="card-body">
                             <div class="form-group">
-                                <label for="inputLanguages">Languages Spoke</label><small class="text-muted m-2">We'll never share your email with anyone else.</small>
+                                <label for="inputLanguages">Languages Spoke</label><small class="text-muted m-2">Select Multiple Languages You Spoked.</small>
                                 <div class="row">
                                     <?php $i = 1;
                                     foreach ($select_lang as $lang_value) {
-                                        echo '<div class="col"><div class="custom-control custom-checkbox">
-                                                <input type="checkbox" name="inputLanguages[]" id="customCheck' . $i . '" class="custom-control-input" value="' . $lang_value . '">
-                                                <label class="custom-control-label" for="customCheck' . $i . '">' . $lang_value . '</label>
-                                            </div></div>';
+                                        echo
+                                            '<div class="col">
+                                            <div class="custom-control custom-checkbox">
+                                            <input type="checkbox" name="inputLanguages[]" id="customCheck' . $i . '" class="custom-control-input ' . $classSpoke . '" value="' . $lang_value . '">
+                                            <label class="custom-control-label" for="customCheck' . $i . '">' . $lang_value . '</label>
+                                        </div></div>';
                                         $i++;
                                     } ?>
                                 </div>
@@ -153,14 +247,16 @@ include(SELECT_HELPER);
                                     <div class="row">
                                         <div class="col">
                                             <div class="custom-control custom-radio custom-control-inline">
-                                                <input type="radio" id="inputGenderMale" name="inputGender" class="custom-control-input" value="male">
+                                                <input type="radio" id="inputGenderMale" name="inputGender" class="custom-control-input <?= $classGender ?>" value="male">
                                                 <label class="custom-control-label" for="inputGenderMale">Male</label>
+                                                <?= $errGender ?>
                                             </div>
                                         </div>
                                         <div class="col">
                                             <div class="custom-control custom-radio custom-control-inline">
-                                                <input type="radio" id="inputGenderFemale" name="inputGender" class="custom-control-input" value="female">
+                                                <input type="radio" id="inputGenderFemale" name="inputGender" class="custom-control-input <?= $classGender ?>" value="female">
                                                 <label class="custom-control-label" for="inputGenderFemale">Female</label>
+                                                <?= $errGender ?>
                                             </div>
                                         </div>
                                     </div>
@@ -169,11 +265,13 @@ include(SELECT_HELPER);
                             <div class="form-row">
                                 <div class="form-group col-md-6">
                                     <label for="inputContactNumber">Contact Number</label>
-                                    <input type="text" name="inputContactNumber" class="form-control" id="inputContactNumber" placeholder="Enter Phone Number">
+                                    <input type="text" name="inputContactNumber" class="form-control <?= $classContact ?>" id="inputContactNumber" placeholder="Enter Phone Number">
+                                    <?= $errContact ?>
                                 </div>
                                 <div class="form-group col-md-6">
                                     <label for="inputEmailAddress">Email Address</label>
-                                    <input type="text" name="inputEmailAddress" class="form-control" id="inputEmailAddress" placeholder="Enter Email Address">
+                                    <input type="text" name="inputEmailAddress" class="form-control <?= $classEmail ?>" id="inputEmailAddress" placeholder="Enter Email Address">
+                                    <?= $errEmail ?>
                                 </div>
                             </div>
                         </div>
@@ -229,58 +327,50 @@ function randomPassword()
 }
 
 if (isset($_POST["savebtn"])) {
-    // $avatar = $conn->real_escape_string($_POST['inputAvatar']);
-    $fname       = $conn->real_escape_string($_POST['inputFirstName']);
-    $lname       = $conn->real_escape_string($_POST['inputLastName']);
-    $speciality = $conn->real_escape_string($_POST['inputSpeciality']);
-    $years      = $conn->real_escape_string($_POST['inputYrsExp']);
-    $desc       = $conn->real_escape_string($_POST['inputDesc']);
-    $lang       = $_POST['inputLanguages'];
-    $spoke      = implode(",", $lang);
-    $dob        = $conn->real_escape_string($_POST['inputDOB']);
-    $gender     = $conn->real_escape_string($_POST['inputGender']);
-    $email      = $conn->real_escape_string($_POST['inputEmailAddress']);
-    $contact    = $conn->real_escape_string($_POST['inputContactNumber']);
+    // ! prefer use empty() for each instead of function multi_empty()  *stackoverflow
+    if (multi_empty($errFName, $errLName, $errSpec, $errYears, $errSpoke, $errGender, $errEmail, $errContact)) {
 
-    if (isset($_FILES["inputAvatar"]["name"])) {
-        $allowed =  array('gif', 'png', 'jpg');
-        $filename = $_FILES['inputAvatar']['name'];
-        $ext = pathinfo($filename, PATHINFO_EXTENSION);
-        if (!in_array($ext, $allowed)) {
-            echo "<script>Swal.fire(
-                'Oops...',
-                'Only can be image!',
-                'error'
-            )</script>";
-            exit();
-        } else {
-            if (!empty($_FILES['inputAvatar']['name'])) {
-                $path = "../database/profile/" . $_FILES['inputAvatar']['name'];
-                $image = $_FILES['inputAvatar']['name'];
-                move_uploaded_file($_FILES['inputAvatar']['tmp_name'], $path);
-            } else {
-                echo "<script>Swal.fire(
-                    'Oops...',
-                    'You should select a file to upload!',
-                    'error'
-                )</script>";
+        if (isset($_FILES["inputAvatar"]["name"])) {
+            $allowed =  array('gif', 'png', 'jpg');
+            $filename = $_FILES['inputAvatar']['name'];
+            $ext = pathinfo($filename, PATHINFO_EXTENSION);
+            if (!in_array($ext, $allowed)) {
+                echo "<script>Swal.fire('Oops...','Only can be image!','error')</script>";
                 exit();
+            } else {
+                if (!empty($_FILES['inputAvatar']['name'])) {
+                    $folderpath = "../uploads/" . $clinic_row['clinic_id'] . "/doctor" . "/";
+                    $path = "../uploads/" . $clinic_row['clinic_id'] . "/doctor" . "/" . $_FILES['inputAvatar']['name'];
+                    $image = $_FILES['inputAvatar']['name'];
+
+                    if (!file_exists($folderpath)) {
+                        mkdir($folderpath, 0777, true);
+                        move_uploaded_file($_FILES['inputAvatar']['tmp_name'], $path);
+                    } else {
+                        move_uploaded_file($_FILES['inputAvatar']['tmp_name'], $path);
+                    }
+                } else {
+                    echo "<script>Swal.fire('Oops...','You should select a file to upload!','error')</script>";
+                    exit();
+                }
             }
         }
-    }
 
-    $password = randomPassword();
+        $password = randomPassword();
 
-    $query = "INSERT INTO doctors (doctor_avatar, doctor_firstname, doctor_lastname, doctor_speciality, doctor_experience, doctor_desc, doctor_password, doctor_spoke, doctor_gender, doctor_dob, doctor_email, doctor_contact, date_created, clinic_id) VALUES ('" . $image . "', '" . $fname . "', '" . $lname . "', '" . $speciality . "', '" . $years . "', '" . $desc . "', '" . $password . "', '" . $spoke . "', '" . $gender . "', '" . $dob . "', '" . $email . "', '" . $contact . "','" . $date_created . "', '" . $clinic_row['clinic_id'] . "')";
-    if (mysqli_query($conn, $query)) {
-        echo '<script>
-            Swal.fire({ "Great!", "New Record Added!", "success" }).then((result) => {
+        $stmt = $conn->prepare("INSERT INTO doctors (doctor_avatar, doctor_firstname, doctor_lastname, doctor_speciality, doctor_experience, doctor_desc, doctor_password, doctor_spoke, doctor_gender, doctor_dob, doctor_email, doctor_contact, date_created, clinic_id) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?)");
+        $stmt->bind_param("ssssissssssssi", $image, $fname, $lname, $speciality, $years, $desc, $password, $spoke, $gender, $dob, $email, $contact, $date_created, $clinic_row['clinic_id']);
+        if ($stmt->execute()) {
+            sendmail($email, $mail['acc_subject'], $mail['acc_title'], $mail['acc_content'], $mail['acc_button'], "/doclab/doctor/verification.php");
+            echo '<script>
+            Swal.fire({ title: "Great!", text: "New Record Added!", type: "success" }).then((result) => {
                 if (result.value) { window.location.href = "doctor-add.php"; }
-            })
+            });
             </script>';
-    } else {
-        echo "Error: " . $query . "<br>" . mysqli_error($conn);
+        } else {
+            echo 'Something Wrong';
+        }
+        $stmt->close();
     }
-    mysqli_close($conn);
 }
 ?>
